+++ packages/poi-detail/src/actions/actions.stories.tsx
@@ -7,7 +7,6 @@ export default {
   component: Actions,
   decorators: [
     (Story) => {
-      localStorage.setItem('REVIEW_TOOLTIP_EXPOSED', 'false')
       return Story()
     },
   ],
