+++ packages/poi-detail/src/actions/actions.tsx
@@ -5,31 +5,19 @@ import {
   HR1,
   Button,
   MarginPadding,
-  Tooltip,
   ButtonGroup,
 } from '@titicaca/core-elements'
-import { useEffect, useRef } from 'react'
-import { useLocalStorage } from '@titicaca/react-hooks'
+import dynamic from 'next/dynamic'
+
+const ReviewTooltip = dynamic(() => import('./review-tooltip'), {
+  ssr: false,
+})
 
 const ActionButton = styled(Button)`
   position: relative;
   padding-left: 0;
   padding-right: 0;
 `
-
-const ReviewTooltip = styled(Tooltip)`
-  width: max-content;
-  padding: 9px 15px 8px;
-  transform: translateX(-50%);
-  left: 50%;
-
-  &::after {
-    transform: translateX(-50%);
-    left: 50%;
-  }
-`
-
-const REVIEW_TOOLTIP_EXPOSED = 'REVIEW_TOOLTIP_EXPOSED'
 
 function Actions({
   scraped,
@@ -53,15 +41,6 @@ function Actions({
   noDivider?: boolean
 }) {
   const { t } = useTranslation('common-web')
-
-  const [isReviewTooltipExposed, setIsReviewTooltipExposed] = useLocalStorage(
-    REVIEW_TOOLTIP_EXPOSED,
-  )
-  const isReviewTooltipExposedCopy = useRef(isReviewTooltipExposed)
-
-  useEffect(() => {
-    setIsReviewTooltipExposed('true')
-  }, [setIsReviewTooltipExposed])
 
   return (
     <Section {...props}>
@@ -92,19 +71,7 @@ function Actions({
           icon={reviewed ? 'starFilled' : 'starEmpty'}
           onClick={onReviewEdit}
         >
-          {isReviewTooltipExposedCopy.current !== 'true' ? (
-            <ReviewTooltip
-              label="이제 영상도 올릴 수 있어요!"
-              pointing={{
-                vertical: 'bottom',
-                horizontal: 'right',
-              }}
-              nowrap={false}
-              borderRadius="16.17"
-              backgroundColor="var(--color-blue)"
-              positioning={{ top: -26 }}
-            />
-          ) : null}
+          <ReviewTooltip />
           {reviewed
             ? t(['ribyusujeong', '리뷰수정'])
             : t(['ribyusseugi', '리뷰쓰기'])}
