+++ packages/poi-detail/src/index.ts
@@ -1,4 +1,4 @@
-export { default as Actions } from './actions'
+export { default as Actions } from './actions/actions'
 export { default as DetailHeader } from './detail-header'
 export { default as ImageCarousel } from './image-carousel'
 export { default as RecommendedArticles } from './recommended-articles'
